refactor(habitat): centralize validation lane routing Moves validation target lane policy into a Habitat validation-routing domain. Hook pre-push planning now consumes that domain plan, Nx target helpers no longer own lane membership, and verify receipt/workspace graph code read verify target names from the domain.
Validation:
- bun run --cwd tools/habitat-harness test -- test/service/hook-service.test.ts test/lib/verify-receipt.test.ts
- bun run --cwd tools/habitat-harness check
- bun tools/habitat-harness/bin/dev.ts check --owner @internal/habitat-harness --json
- bun run openspec -- validate deep-habitat-effect-validation-lane-routing --strict
- bun run biome:ci
- git diff --check
